# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Apigo - Spec Documentation API
  version: "1.0"
paths:
...
  /user/profile/:id:
    get:
      consumes:
        - application/json
      description: user update
      parameters:
        - description: Bearer
          in: header
          name: Authorization
          required: true
          type: string
        - description: userId
          format: userId
          in: query
          name: userId
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: user account
      tags:
        - User Profile
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"